requirements.txt
pyyaml
numpy
pandas
scipy
matplotlib
opencv-python
pillow
torch
torchvision
openpifpaf
